requirements.txt
Adafruit-Blinka==8.0.2
adafruit-circuitpython-busdevice==5.1.10
adafruit-circuitpython-typing==1.7.1
Adafruit-PlatformDetect==3.24.1
Adafruit-PureIO==1.1.9
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
idna==3.3
imageio==2.19.2
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.2
notebook==6.4.11
numpy==1.22.4
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pipreqs==0.4.11
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyftdi==0.54.0
pygame==2.1.2
Pygments==2.12.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2022.1
pyusb==1.2.1
PyWavelets==1.3.0
pyzmq==23.0.0
requests==2.27.1
scikit-image==0.19.2
scikit-learn==1.1.1
scipy==1.8.1
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
terminado==0.15.0
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
tornado==6.1
traitlets==5.2.1.post0
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
yarg==0.1.9
